Changed int to long

+++ src/main/java/com/odeyalo/sonata/authorization/dto/TokenDto.java
@@ -11,5 +11,5 @@ import lombok.experimental.FieldDefaults;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class TokenDto {
     String value;
-    int lifetime;
+    long lifetime;
 }
